backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
celery==5.3.4
redis==5.0.1
supabase==2.0.1
langchain==0.1.0
langgraph==0.0.25
deepgram-sdk==3.2.0
elevenlabs==0.2.26
mistral-common==0.0.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
email-validator==2.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
pillow==10.1.0
PyPDF2==3.0.1
requests==2.31.0
stripe==7.8.0
slack-sdk==3.26.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
python-socketio==5.10.0
weaviate-client==4.1.1
novu===1.1.0
